Fix stale audio text dimming
